fix: move Python GrowableBuffer to numba module and check and initialize Numba when it is constructed
